% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_py_packs}
\alias{install_py_packs}
\title{Setup python packages}
\usage{
install_py_packs(envname = NULL, method = "auto", conda = "auto")
}
\arguments{
\item{envname}{The name, or full path, of the environment in which Python
packages are to be installed. When NULL (the default), the active
environment as set by the RETICULATE_PYTHON_ENV variable will be used; if
that is unset, then the r-reticulate environment will be used.}

\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate
to automatically find an appropriate conda binary. See Finding Conda and
conda_binary() for more details.}
}
\description{
Setup python packages
}
